Moved file constructors and add_input/output to public API.
